requirements.txt
# Web框架
fastapi==0.135.1
uvicorn==0.42.0
# 文件处理
pdf2docx==0.5.12
tabula-py==2.10.0
pdfplumber==0.11.9
openpyxl==3.1.5
PyMuPDF>=1.26.7
python-docx==1.1.2
rapidocr-onnxruntime==1.2.3
# 其他
python-multipart==0.0.22  # 处理文件上传
APScheduler==3.11.2      # 定时任务
slowapi==0.1.9           # 限流
limits==5.8.0            # 限流依赖
python-dotenv==1.2.2      # 加载.env文件
starlette==1.0.0
pydantic==2.12.5
pydantic-settings==2.14.0